LayoutTests/imported/w3c/web-platform-tests/workers/modules/dedicated-worker-parse-error-failure-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token ';'
CONSOLE MESSAGE: SyntaxError: Unexpected token ';'

PASS Module worker construction for script with syntax error should dispatch an event named error.
PASS Static import on module worker for script with syntax error should dispatch an event named error.
